Start consolidating shared types.
